+++ builder/src/main/scala/com/github/gchudnov/bscript/builder/AstMeta.scala
@@ -1,7 +1,7 @@
 package com.github.gchudnov.bscript.builder
 
 import com.github.gchudnov.bscript.lang.ast.AST
-import com.github.gchudnov.bscript.builder.state.Meta
+import com.github.gchudnov.bscript.builder.Meta
 
 final case class AstMeta(
   ast: AST,
