+++ builder/src/main/scala/com/github/gchudnov/bscript/builder/Scope.scala
@@ -1,4 +1,4 @@
-package com.github.gchudnov.bscript.builder.state
+package com.github.gchudnov.bscript.builder
 
 trait Scope:
   def name: String
