+++ builder/src/main/scala/com/github/gchudnov/bscript/builder/internal/ScopeBuilder.scala
@@ -1,6 +1,7 @@
-package com.github.gchudnov.bscript.builder.state
+package com.github.gchudnov.bscript.builder
 
 import com.github.gchudnov.bscript.builder.util.Forest
+import com.github.gchudnov.bscript.builder.Meta
 
 /**
  * Scope
@@ -40,26 +41,11 @@ trait ScopeBuilder:
 
   def define(s: Symbol): Unit
 
+  def result: Meta
+
 /**
  * ScopeBuilder
  */
 object ScopeBuilder:
 
   type ScopeTree = Forest[Scope]
-
-/**
- * BasicScopeBuilder
- */
-final class BasicScopeBuilder() extends ScopeBuilder:
-  import ScopeBuilder.*
-
-  private var scopeTree: ScopeTree = Forest.empty[Scope]
-
-  override def push(): Unit =
-    ???
-
-  override def pop(): Unit =
-    ???
-
-  override def define(s: Symbol): Unit =
-    ???
