+++ builder/src/main/scala/com/github/gchudnov/bscript/builder/internal/TypeCheckVisitor.scala
@@ -3,7 +3,8 @@ package com.github.gchudnov.bscript.builder.internal
 import com.github.gchudnov.bscript.builder.TypeCheckLaws
 import com.github.gchudnov.bscript.builder.TypeCheckLaws.*
 // import com.github.gchudnov.bscript.builder.internal.TypeCheckVisitor.*
-import com.github.gchudnov.bscript.builder.state.{ Ctx, Meta }
+import com.github.gchudnov.bscript.builder.state.{ Ctx }
+import com.github.gchudnov.bscript.builder.{ Meta }
 import com.github.gchudnov.bscript.lang.ast.*
 import com.github.gchudnov.bscript.lang.symbols.*
 import com.github.gchudnov.bscript.lang.util.{ Casting, Transform }
