+++ builder/src/test/scala/com/github/gchudnov/bscript/builder/internal/TypeCheckVisitorSpec.scala
@@ -6,13 +6,11 @@ import com.github.gchudnov.bscript.lang.ast.*
 // import com.github.gchudnov.bscript.builder.internal.TypeCheckVisitor.TypeCheckState
 // import com.github.gchudnov.bscript.builder.BGlobals
 import com.github.gchudnov.bscript.builder.BTypeCheckLaws
-import com.github.gchudnov.bscript.lang.ast.CompiledExpr
 import com.github.gchudnov.bscript.lang.symbols.*
-import com.github.gchudnov.bscript.builder.state.Meta
+import com.github.gchudnov.bscript.builder.Meta
 import com.github.gchudnov.bscript.lang.types.{ TypeNames }
 import com.github.gchudnov.bscript.lang.util.Transform
 import com.github.gchudnov.bscript.builder.util.Gen
-import com.github.gchudnov.bscript.builder.state.Meta
 import com.github.gchudnov.bscript.builder.internal.MetaOps
 import com.github.gchudnov.bscript.builder.TestSpec
